requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
discord.py==2.6.4
dotenv==0.0.5
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.7
python-dotenv==1.1.1
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
